[DTensor] Fix compute_local_shape_and_global_offset for uneven sharding
This fix is needed for cases where distributed checkpointing (DCP) is used to
save a local state dict.  That's becuase DCP relies on the local-shape / global-offset
for each rank being correct to save files that can be correctly
resharded (or indeed loaded at all).

(If saving a 'full state dict' instead of a local one, DCP would convert
to 'full tensors' before saving, and that logic got fixed in the
previous PR in this stack.)

Also add a util `_explicit_order_placements` which converts a list of
placements with StridedSharding into a list of placements with only
regular sharding, with the order shuffled such that it is equivalent.

ghstack-source-id: 967a1118e090c687356f5ff7e23b050f507c0895
Pull Request resolved: https://github.com/pytorch/pytorch/pull/150393